requirements.txt
logging
streamlit
pdfplumber
tabula-py
json
nltk